requirements/common.txt
alembic==1.8.1
async-generator==1.10
attrs==22.1.0
bleach==5.0.1
blinker==1.5
certifi==2022.9.24
click==8.1.3
coverage==6.5.0
dnspython==2.2.1
dominate==2.7.0
email-validator==1.3.0
Flask==2.2.2
Flask-Bootstrap==3.3.7.1
Flask-HTTPAuth==4.7.0
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-Moment==1.0.5
Flask-PageDown==0.4.0
flask-sock==0.5.2
Flask-SQLAlchemy==3.0.2
Flask-SSLify==0.1.5
Flask-WTF==1.0.1
greenlet==1.1.3.post0
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.3
Markdown==3.4.1
MarkupSafe==2.1.1
outcome==1.2.0
packaging==21.3
parso==0.8.3
PyJWT==2.6.0
PyMySQL==1.0.2
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
simple-websocket==0.9.0
six==1.16.0
sniffio==1.3.0
sock==0.6.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.42
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.13
visitor==0.1.3
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==2.2.2
wsproto==1.2.0
WTForms==3.0.1
